renamed motion manager
