+++ KF_Framework.vcxproj
@@ -157,7 +157,7 @@
     <ClCompile Include="damagedState.cpp" />
     <ClCompile Include="jumpState.cpp" />
     <ClCompile Include="landState.cpp" />
-    <ClCompile Include="motionManager.cpp" />
+    <ClCompile Include="motion_manager.cpp" />
     <ClCompile Include="neutralState.cpp" />
     <ClCompile Include="UIObject2D.cpp" />
     <ClCompile Include="rigidbody3D.cpp" />
@@ -226,7 +226,7 @@
     <ClInclude Include="jumpState.h" />
     <ClInclude Include="landState.h" />
     <ClInclude Include="light.h" />
-    <ClInclude Include="motionManager.h" />
+    <ClInclude Include="motion_manager.h" />
     <ClInclude Include="actorParamater.h" />
     <ClInclude Include="neutralState.h" />
     <ClInclude Include="shader_enum.h" />
